transfer code to delegate
