+++ app/src/main/res/layout/run_code_spinner_item.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <TextView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@android:id/text1"
-    android:padding="8dp"
+    android:padding="12dp"
     android:textSize="16sp"
     android:textColor="@color/violet1"
     android:singleLine="true"
